+++ x/bank/keeper/send.go
@@ -13,6 +13,7 @@ import (
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 	"github.com/cosmos/cosmos-sdk/x/bank/types"
 	paramtypes "github.com/cosmos/cosmos-sdk/x/params/types"
+	"github.com/tendermint/tendermint/libs/math"
 )
 
 // SendKeeper defines a module interface that facilitates the transfer of coins
@@ -24,7 +25,7 @@ type SendKeeper interface {
 	PrependSendRestriction(restriction types.SendRestrictionFn)
 	ClearSendRestriction()
 
-	InputOutputCoins(ctx sdk.Context, input types.Input, outputs []types.Output) error
+	InputOutputCoins(ctx sdk.Context, inputs []types.Input, outputs []types.Output) error
 	SendCoins(ctx sdk.Context, fromAddr, toAddr sdk.AccAddress, amt sdk.Coins) error
 
 	GetParams(ctx sdk.Context) types.Params
@@ -126,45 +127,75 @@ func (k BaseSendKeeper) SetParams(ctx sdk.Context, params types.Params) {
 	k.paramSpace.SetParamSet(ctx, &p)
 }
 
-// InputOutputCoins performs multi-send functionality. It accepts an
-// input that corresponds to a series of outputs. It returns an error if the
-// input and outputs don't line up or if any single transfer of tokens fails.
-func (k BaseSendKeeper) InputOutputCoins(ctx sdk.Context, input types.Input, outputs []types.Output) error {
-	// Safety check ensuring that when sending coins the keeper must maintain the
-	// Check supply invariant and validity of Coins.
-	if err := types.ValidateInputsOutputs(input, outputs); err != nil {
+// InputOutputCoins performs multi-send functionality. There must be exactly one input and/or exactly one output.
+// It returns an error if the input and outputs don't line up or if any single transfer of tokens fails.
+func (k BaseSendKeeper) InputOutputCoins(ctx sdk.Context, inputs []types.Input, outputs []types.Output) error {
+	// As a keeper function, we can't assume that MsgMultiSend.ValidateBasic was called on these inputs and outputs.
+	if err := types.ValidateInputsOutputs(inputs, outputs); err != nil {
 		return err
 	}
 
-	inAddress, err := sdk.AccAddressFromBech32(input.Address)
-	if err != nil {
-		return err
+	// Remove the funds from the inputs first as that's the most common point of failure.
+	for i, input := range inputs {
+		inAddress := sdk.MustAccAddressFromBech32(input.Address)
+		err := k.subUnlockedCoins(ctx, inAddress, input.Coins)
+		if err != nil {
+			return fmt.Errorf("input[%d] %s: %w", i, input.Address, err)
+		}
+		ctx.EventManager().EmitEvent(
+			sdk.NewEvent(
+				sdk.EventTypeMessage,
+				sdk.NewAttribute(types.AttributeKeySender, input.Address),
+			),
+		)
 	}
 
-	err = k.subUnlockedCoins(ctx, inAddress, input.Coins)
-	if err != nil {
-		return err
-	}
-
-	ctx.EventManager().EmitEvent(
-		sdk.NewEvent(
-			sdk.EventTypeMessage,
-			sdk.NewAttribute(types.AttributeKeySender, input.Address),
-		),
-	)
-
-	for _, out := range outputs {
-		outAddress, err := sdk.AccAddressFromBech32(out.Address)
+	// Create a map of AccAddress (cast to string) to the amount that that address will get.
+	// The keys are the addresses that come back from the send restriction, not necessarily the addresses in the outputs.
+	// Keep track of the order of the output address too since looping over a map is non-deterministic.
+	toOutput := make(map[string]sdk.Coins)
+	outputOrder := make([]sdk.AccAddress, 0, math.MaxInt(len(inputs), len(outputs)))
+	// applySendRestriction will make the call to the send restriction function,
+	// and update the toOutput and outputOrder values accordingly.
+	applySendRestriction := func(inputAddress, outputAddress string, coins sdk.Coins) error {
+		inAddr := sdk.MustAccAddressFromBech32(inputAddress)
+		outAddrOrig := sdk.MustAccAddressFromBech32(outputAddress)
+		outAddr, err := k.sendRestriction.apply(ctx, inAddr, outAddrOrig, coins)
 		if err != nil {
 			return err
 		}
-
-		outAddress, err = k.sendRestriction.apply(ctx, inAddress, outAddress, out.Coins)
-		if err != nil {
-			return err
+		amt, known := toOutput[string(outAddr)]
+		if !known {
+			outputOrder = append(outputOrder, outAddr)
 		}
+		toOutput[string(outAddr)] = amt.Add(coins...)
+		return nil
+	}
 
-		if err = k.addCoins(ctx, outAddress, out.Coins); err != nil {
+	// If there's multiple inputs, we apply the send restriction for each input.
+	// Otherwise, apply the send restriction for each output.
+	// ValidateInputsOutputs prevents many-to-many.
+	if len(inputs) > 1 {
+		for _, input := range inputs {
+			err := applySendRestriction(input.Address, outputs[0].Address, input.Coins)
+			if err != nil {
+				return err
+			}
+		}
+	} else {
+		for _, output := range outputs {
+			err := applySendRestriction(inputs[0].Address, output.Address, output.Coins)
+			if err != nil {
+				return err
+			}
+		}
+	}
+
+	// Finally, add the coins to the appropriate account(s).
+	for _, outAddress := range outputOrder {
+		amt := toOutput[string(outAddress)]
+		err := k.addCoins(ctx, outAddress, amt)
+		if err != nil {
 			return err
 		}
 
@@ -172,7 +203,7 @@ func (k BaseSendKeeper) InputOutputCoins(ctx sdk.Context, input types.Input, out
 			sdk.NewEvent(
 				types.EventTypeTransfer,
 				sdk.NewAttribute(types.AttributeKeyRecipient, outAddress.String()),
-				sdk.NewAttribute(sdk.AttributeKeyAmount, out.Coins.String()),
+				sdk.NewAttribute(sdk.AttributeKeyAmount, amt.String()),
 			),
 		)