+++ CHANGELOG.md
@@ -44,6 +44,14 @@ Ref: https://keepachangelog.com/en/1.0.0/
 * [#578](https://github.com/provenance-io/cosmos-sdk/pull/578) Add `binary_version` to the `NodeInfo` object returned by status command.
 * [#577](https://github.com/provenance-io/cosmos-sdk/pull/577) Add an injectable `GetLockedCoinsFn` to the bank module.
 
+### Improvements
+
+* [#581](https://github.com/provenance-io/cosmos-sdk/pull/581) For `MsgMultiSend` and `InputOutputCoins`, allow many inputs when there's a single output.
+
+### API Breaking
+
+* [#581](https://github.com/provenance-io/cosmos-sdk/pull/581) The `InputOutputCoins` once again takes in multiple inputs. It returns an error if there are multiple inputs and multiple outputs.
+
 ---
 
 ## [v0.46.13-pio-1](https://github.com/provenance-io/cosmos-sdk/releases/tag/v0.46.13-pio-1) - 2023-06-09